Example file

+++ src/main/java/com/spotinst/sdkjava/model/api/azure/statefulNode/ApiStatefulNodeSignalConfiguration.java
@@ -21,7 +21,7 @@ public class ApiStatefulNodeSignalConfiguration implements IPartialUpdateEntity 
     private Set<String> isSet;
     private String      type;
     private Integer     timeout;
-    private Long        enabledAt;
+    private Integer     enabledAt;
 
     public ApiStatefulNodeSignalConfiguration() {
         isSet = new HashSet<>();
@@ -53,11 +53,11 @@ public class ApiStatefulNodeSignalConfiguration implements IPartialUpdateEntity 
         this.timeout = timeout;
     }
 
-    public Long getEnabledAt() {
+    public Integer getEnabledAt() {
         return enabledAt;
     }
 
-    public void setEnabledAt(Long enabledAt) {
+    public void setEnabledAt(Integer enabledAt) {
         isSet.add("enabledAt");
         this.enabledAt = enabledAt;
     }
